Ensure median_ci indices are int, not float

+++ experiments/utils.py
@@ -51,8 +51,8 @@ def median_ci(values, confidence_level=0.95):
     z_factor = norm().ppf(1 - critical_value / 2)
 
     # Compute CI rankings
-    low_ci_rank = round(num_repeats / 2 - z_factor * np.sqrt(num_repeats) / 2)
-    high_ci_rank = round(1 + num_repeats / 2 + z_factor * np.sqrt(num_repeats) / 2)
+    low_ci_rank = int(round(num_repeats / 2 - z_factor * np.sqrt(num_repeats) / 2))
+    high_ci_rank = int(round(1 + num_repeats / 2 + z_factor * np.sqrt(num_repeats) / 2))
     values_sorted = sorted(values)
 
     return values_sorted[low_ci_rank], median, values_sorted[high_ci_rank]
